requirements.txt
experta==1.9.4
streamlit==1.28.0
pandas==2.0.3
groq==0.4.1
python-dotenv==1.0.0